more improvements

+++ Engine/Engine/Camera.h
@@ -32,9 +32,9 @@ class Camera: public Object{
 
 		virtual void update(float);
 
-		void lookAt(glm::vec3);  
-		void lookAt(glm::vec3,glm::vec3); 
-		void lookAt(glm::vec3,glm::vec3,glm::vec3); 
+		void lookAt(glm::dvec3);  
+		void lookAt(glm::dvec3,glm::dvec3); 
+		void lookAt(glm::dvec3,glm::dvec3,glm::dvec3); 
 		void lookAt(Object*,bool targetUp = false);
 
 		void setAspectRatio(float);
@@ -54,7 +54,7 @@ class Camera: public Object{
 		const CAMERA_TYPE getType() const{ return m_Type; }
 
 		bool sphereIntersectTest(ObjectDisplay*);
-		bool sphereIntersectTest(glm::vec3 pos, float radius);
+		bool sphereIntersectTest(glm::dvec3 pos, float radius);
 
 		//ray tests
 		bool rayIntersectSphere(ObjectDisplay*);
